# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewPlate API - Smart NFC/QR review card platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-profiles:
    get:
      operationId: listBusinessProfiles
      tags: [business-profiles]
      summary: List business profiles for current user
      responses:
        "200":
          description: List of business profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessProfile"

    post:
      operationId: createBusinessProfile
      tags: [business-profiles]
      summary: Create a business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "400":
          description: Plan limit reached or bad data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business-profiles/{id}:
    get:
      operationId: getBusinessProfile
      tags: [business-profiles]
      summary: Get a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateBusinessProfile
      tags: [business-profiles]
      summary: Update a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"

    delete:
      operationId: deleteBusinessProfile
      tags: [business-profiles]
      summary: Delete a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CARDS
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List cards for current user
      parameters:
        - name: businessProfileId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardBody"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/activate:
    post:
      operationId: activateCard
      tags: [cards]
      summary: Activate a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/deactivate:
    post:
      operationId: deactivateCard
      tags: [cards]
      summary: Deactivate a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/activate-by-code:
    post:
      operationId: activateCardByCode
      tags: [cards]
      summary: Activate a card by its unique code (user claims the card)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateByCodeBody"
      responses:
        "200":
          description: Card activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SCAN LOGS
...
  /public/scan:
    get:
      operationId: publicScan
      tags: [public]
      summary: Public scan redirect - logs scan and returns redirect URL
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan result with redirect URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /public/card/{code}:
    get:
      operationId: getPublicCard
      tags: [public]
      summary: Get public card data by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public card data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ADMIN
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        customBannerText:
          type: ["string", "null"]
        customBannerColor:
          type: ["string", "null"]
        showPoweredBy:
          type: boolean
        cardCount:
          type: integer
        totalScans:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - name
        - showPoweredBy
        - cardCount
        - totalScans
        - createdAt
    CreateBusinessProfileBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
      required:
        - name
    UpdateBusinessProfileBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        customBannerText:
          type: ["string", "null"]
        customBannerColor:
          type: ["string", "null"]
        showPoweredBy:
          type: ["boolean", "null"]
    Card:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        status:
          type: string
        platform:
          type: ["string", "null"]
        targetUrl:
          type: ["string", "null"]
        businessProfileId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        scanCount:
          type: integer
        smartReviewEnabled:
          type: boolean
        negativeAlertEnabled:
          type: boolean
        createdAt:
          type: string
        activatedAt:
          type: ["string", "null"]
      required:
        - id
        - code
        - status
        - scanCount
        - smartReviewEnabled
        - negativeAlertEnabled
        - createdAt
    UpdateCardBody:
      type: object
      properties:
        platform:
          type: ["string", "null"]
        targetUrl:
          type: ["string", "null"]
        businessProfileId:
          type: ["integer", "null"]
        smartReviewEnabled:
          type: ["boolean", "null"]
        negativeAlertEnabled:
          type: ["boolean", "null"]
    ActivateByCodeBody:
      type: object
      properties:
        code:
          type: string
        businessProfileId:
          type: integer
      required:
        - code
        - businessProfileId
...
    ScanResult:
      type: object
      properties:
        redirectUrl:
          type: string
        platform:
          type: string
        smartReviewEnabled:
          type: boolean
        cardCode:
          type: string
      required:
        - redirectUrl
        - platform
        - smartReviewEnabled
        - cardCode
    PublicCard:
      type: object
      properties:
        code:
          type: string
        platform:
          type: string
        targetUrl:
          type: string
        businessName:
          type: ["string", "null"]
        businessLogoUrl:
          type: ["string", "null"]
        smartReviewEnabled:
          type: boolean
        showPoweredBy:
          type: boolean
        customBannerText:
          type: ["string", "null"]
        customBannerColor:
          type: ["string", "null"]
      required:
        - code
        - platform
        - targetUrl
        - smartReviewEnabled
        - showPoweredBy
...